requirements.txt
langchain>=0.1.14
pypdf2>=3.0.1
python-docx>=1.1.0
openpyxl>=3.1.2
pandas>=2.2.1
llama-index>=0.10.20  # Optional, if you plan to integrate local LLMs
faiss-cpu>=1.8.0      # Optional, for similarity search if added later
